Implemented FiniteStateMachine events - FiniteStateMachine now calls events when exiting AND entering a new state. - FiniteStateMachine now has a OnUpdateStateParams struct for dealing with event dispatchers. - Removed some unneeded casting while getting the window size - Some more cleaning up as usaul

+++ Plasma/Engine/Components/SpriteComponent.cpp
@@ -62,7 +62,7 @@ void SpriteComponent::CalcBounds()
 SpriteComponent* SpriteComponent::Set(Texture2D sprite) { this->sprite = sprite; return this; }
 Texture2D SpriteComponent::Get() { return this->sprite; }
 
-SpriteComponent* SpriteComponent::AnimationSpeed(double speed) { this->animationSpeed = speed; return this; }
+SpriteComponent* SpriteComponent::AnimationSpeed(int speed) { this->animationSpeed = speed; return this; }
 SpriteComponent* SpriteComponent::SetColour(Colour newColour) { color = newColour; return this; }
 int SpriteComponent::GetAnimationSpeed() { return this->animationSpeed; }
